pyproject.toml
[project]
name = "langgraph-app-astradb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "cassio>=0.1.10",
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-astradb>=0.6.0",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.59",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.1.2",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.2",
    "langsmith>=0.3.42",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
    "validators>=0.35.0",
    "wikipedia>=1.4.0",
]
